Revert "fix:battle heal dead target"
This reverts commit 1b5544b86745d69baca4110560fe04afe1cf3c11.

+++ SaSH/script/action.cpp
@@ -318,8 +318,7 @@ qint64 Interpreter::dropitem(qint64, const TokenMap& TK)
 	if (tempName.isEmpty() && memo.isEmpty() && itemIndex == -1)
 		return Parser::kArgError + 1ll;
 	else if (tempName.isEmpty() && memo.isEmpty()
-		&& ((itemIndex >= 1 && itemIndex <= static_cast<qint64>(MAX_ITEM - CHAR_EQUIPPLACENUM))
-			|| (itemIndex >= 101 && itemIndex <= static_cast<qint64>(CHAR_EQUIPPLACENUM + 100))))
+		&& ((itemIndex >= 1 && itemIndex <= (MAX_ITEM - CHAR_EQUIPPLACENUM)) || (itemIndex >= 101 && itemIndex <= (CHAR_EQUIPPLACENUM + 100))))
 	{
 		if (itemIndex < 100)
 			--itemIndex;
@@ -1225,7 +1224,7 @@ qint64 Interpreter::petunequip(qint64, const TokenMap& TK)
 
 	qint64 petIndex = -1;
 	if (!checkInteger(TK, 1, &petIndex))
-		return Parser::kArgError + 1ll;
+		return Parser::kArgError + 1l;
 
 	--petIndex;
